Fix bug #2224

+++ lib/MySQL_Session.cpp
@@ -3401,7 +3401,7 @@ handler_again:
 			}
 			if (mysql_thread___connect_timeout_server_max) {
 				if (mybe->server_myds->max_connect_time==0)
-					mybe->server_myds->max_connect_time=thread->curtime+mysql_thread___connect_timeout_server_max*1000;
+					mybe->server_myds->max_connect_time=thread->curtime+(long long)mysql_thread___connect_timeout_server_max*1000;
 			} else {
 				mybe->server_myds->max_connect_time=0;
 			}
